+++ tutorial_2/src/softmax_restrict.cpp
@@ -36,7 +36,7 @@ void softmax_scale(float* __restrict__ output,
 
     float sum = 0.0f;
     for (int i = 0; i < N; ++i) {
-        output[i] = std::expf(input[i] - max_val);
+        output[i] = std::exp(input[i] - max_val);
         sum += output[i];
     }
